- fix: PostComp.CommentCount 이동 및 자잘한 변경사항.
